requirements.txt
Flask==2.0.3
tensorflow==2.18.0
numpy==1.19.5
Pillow==8.4.0
werkzeug==2.0.3